change floatfield to decimalfield

+++ src/main/models.py
@@ -76,7 +76,7 @@ class Country(models.Model):
     place = models.IntegerField(_('Place on rating'), default=1)
     points = models.IntegerField(_('Points on rating'), default=1)
     place_ass = models.IntegerField(_('Place on association rating'), default=1)
-    points_ass = models.IntegerField(_('Points on association rating'), default=1)
+    points_ass = models.DecimalField(_('Points on association rating'), max_digits=6, decimal_places=3, default=1)
     cl_teams = models.IntegerField(_('The number of represented in the Champions League'), default=0)
     el_teams = models.IntegerField(_('The number of represented in the Europe League'), default=1)
     wc = models.IntegerField(_('World Cup'), default=0)
@@ -275,7 +275,7 @@ class Club(models.Model):
     manager = models.CharField(_('Manager'), null=True, max_length=255)
     website = models.URLField(_('Website'), null=True, max_length=255)
     place = models.IntegerField(_('Place on rating'), default=1)
-    points = models.FloatField(_('Points on rating'), default=1.000)
+    points = models.DecimalField(_('Points on rating'), max_digits=6, decimal_places=3, default=1)
     cups = models.ManyToManyField(
         'Cup', related_name='club',
         verbose_name=_('cup'), blank=True,
